fix: minor errors
